requirements.txt
opencv-python
tslearn
networkx
oss2
pillow
pandas
loguru